pcie: Handle PERST signal from the gpio.
This patch will use the two gpio to act as EP-PERST(input) and
RC-PERST(output).
EP-PERST(input): When this gpio sense the GPIO falling edge irq the
patch will execute the below procedure to fix the EP/RC mctp issue:
1. Remove the devices on the RC bus
2. Toggle the SSPRST# to reset the PCIe EP and PCIe RC
3. Rescan the devices on the RC bus
RC-PERST(output): When the BMC reboot the patch will toggle the gpio to
reset the devices on it.
In addition, this patch will remove the toggle of pcie rc perst from the
pciecfg driver to avoid reset the EP when BMC reboot.

The topology for MCTP EP/RC work at the same time:
+------+            +---------------+
|x86 RC+----------->|GPIOx(3.3v)    |
+------+            |               |
         +----------+PERST#         |
         |          |               |
         |          |               |
         +--------->|SSPRST#        |
+------+            |               |
|  EP  |<-----------+GPIOy(3.3v)    |
+------+            |               |
                    +---------------+
dts setting: GPIOx = GPIOB0, GPIOy = GPIOB1 as sample
&pciecfg {
	pcei1-perst-gpios = <&gpio0 ASPEED_GPIO(B, 1) GPIO_ACTIVE_HIGH>;
	...
};

&pcie1 {
	perst-ep-in-gpios = <&gpio0 ASPEED_GPIO(B, 0) GPIO_ACTIVE_HIGH>;
	perst-rc-out-gpios = <&gpio0 ASPEED_GPIO(B, 1) GPIO_ACTIVE_HIGH>;
	...
};

Signed-off-by: ryan_chen <ryan_chen@aspeedtech.com>
Change-Id: I1663c40228e1d8481ddc9268800535f45994efc5