media: aspeed: Fix kernel panic at probe in 2500
Kernel panic will be caused by null pointer used in regmap_read in
aspeed_video_decide_input(). This is because we only support GFX/MEM
input in 2600, so regmap for gfx/scu will be null in other chips.

Assign ERR_PTR to regmap of gfx/scu to fix the problem.

Signed-off-by: Jammy Huang <jammy_huang@aspeedtech.com>
Change-Id: I82060f13ddd2f19dbb1d05114f034fc83420ffcf