i3c: add i3c eeprom sample driver for slave mode
Add a sample driver to simulate EEPROM in I3C slave mode. Due to I3C's
design, it is impossible to prepare the read data as soon as the I3C
slave controller receiving read command.  So the pending read
notification is used to notify the main master that the read data is
ready to be read.  The sequence is as:

1. master sends a write transfer.  The first byte of the transfer
specifies the address offset of the EEPROM.  And the following bytes are
write data bytes.

2. slave does the write operation.  Then fill the read data to the HW TX
queue and issues IBI.

3. master decodes the IBI and issues a read transfer to read the data
back.

Signed-off-by: Dylan Hung <dylan_hung@aspeedtech.com>
Change-Id: I842f37da0e015e5f1631ec6f7ab03ebc2e4b763f
(cherry picked from commit aa9a4b90619201a1e4c8cdf4c30e67c4de7da82e)