pcie: Add the GPIO to control the ownership of the PERST.
This patch add the GPIO to change the extenal mux which is used to
change the signal owner of the PCIe EP PERST#. Before the GPIOx can
handle the PERST# the owner of the signal should be the host to avoid
the BMC PCIe EP missing the perst reset. After that the owner of the
signal should be the SSPRST# to avoid host power on/off affecting the
function of the MCTP over PCIe RC port.
The dts setting for this GPIO sholud like below:
&pcie1 {
    ...
    perst-owner-gpios = <&gpio0 ASPEED_GPIO(B, 2)
			(GPIO_ACTIVE_HIGH | GPIO_TRANSITORY)>;
    ...
};
GPIO_TRANSITORY is used to auto clear the gpio setting after WDT reset.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: I21af166100b19795884c8f9f60c49d04c97bad94