i3c: master: Restore more information of olddev.
When the bus find the device with the same pid, the i3c subsystem will
treat it as the same. The old one will be removed from the bus and
restore the status to the new one. This patch adds items that need to be
restored: mrl/mwl/ibi payload length and the pec enable.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: If2f4c42774174c5233215cc55a86678520a89345