i3c: ast2600: Add API to toggle the scl in by software mode
Before the reset FIFO,the i3c controller should be disabled to avoid the
incomplete resetting. After that the i3c controller should be enabled to
avoid the next transfer to fail.
The enable bit of the i3c slave mode require the scl to active, but we
can't control the scl in slave mode. So, this patch add the api to use
the software force mode to generate the dummy scl_in signal which won't
affect external signal to active the bit.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: Idebd9ae2745927ab2db294a5ff7554685857a8a2
(cherry picked from commit c1a835c4db469ea6356af0226bfc3002d870efb5)