pinctrl: i3c: Add dependency for register setting to avoid abnormal behavior
The I3C3/4 controllers have the capability to configure the function to
operate at either high or low voltage pins depending on specific
scenarios. However, simultaneous activation of both voltage pins should
be avoided to prevent undesired behaviors in the I3C.
This patch introduces a dependency for the register setting to ensure that
such a situation is avoided.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: I509f2b6ea1b966ad696134cd68a937d39e49fd37