i3c: aspeed: Active the enable bit by internal operations
The enable bit of the i3c slave mode require the scl and stop sequences
to active, but we can't control the scl in slave mode. So, this patch add
support for toggling SCL and generating stop sequences for internal bus
operations which won't affect external signal to active the bit.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: Ia6c00c7445be3e0708f00c1093ac4710dff6857f