soc: aspeed: mctp: Fix the condition for clearing rx_warmup.
The rx_warmup flag should be cleared after the first loop when the chip
version is A1/A2. This is because the initial loop of the receive buffer
might conclude from an unpredictable location. Hence, the driver must
continue searching for the packet in all buffers until the wrap-around
condition stabilizes after the initial loop.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: Id1c63b92d21a0f00989cc89aa8cb1cbc2baf6872